python-whatsapp-milvus/requirements.txt
flask
twilio
pymilvus
weasyprint
python-dotenv
requests
datetime
uuid
openai==0.28
pydub
gunicorn
google.generativeai